+++ src/lib.rs
@@ -90,7 +90,7 @@
 //!
 //! ## Owned, aka Consuming
 //!
-//! Precede your struct with `#[setters(owned)]` to opt into this pattern.
+//! Precede your struct (or field) with `#[setter(owned)]` to opt into this pattern.
 //!
 //! * Setters take and return `self`.
 //! * PRO: Setter calls and final build method can be chained.
@@ -100,7 +100,7 @@
 //! ## Mutable, aka Non-Comsuming (recommended)
 //!
 //! This pattern is recommended and active by default if you don't specify anything else.
-//! You can precede your struct with `#[setters(mutable)]` to make this choice explicit.
+//! You can precede your struct (or field) with `#[setter(mutable)]` to make this choice explicit.
 //!
 //! * Setters take and return `&mut self`.
 //! * PRO: Setter calls and final build method can be chained.
@@ -109,7 +109,7 @@
 //!
 //! ## Immutable
 //!
-//! Precede your struct with `#[setters(immutable)]` to opt into this pattern.
+//! Precede your struct (or field) with `#[setter(immutable)]` to opt into this pattern.
 //!
 //! * Setters take and return `&self`.
 //! * PRO: Setter calls and final build method can be chained.
@@ -177,10 +177,18 @@
 //!
 //! ## Setter Visibility
 //!
-//! Setters are public by default. You can precede your struct with `#[setters(public)]`
+//! Setters are public by default. You can precede your struct (or field) with `#[setter(public)]`
 //! to make this explicit.
 //!
-//! Otherwise precede your struct with `#[setters(private)]` to opt into private setters.
+//! Otherwise precede your struct (or field) with `#[setter(private)]` to opt into private setters.
+//!
+//! ## Setter Prefixes
+//!
+//! Setter methods are named after their corresponding field by default.
+//!
+//! You can precede your struct (or field) with e.g. `#[setter(prefix="xyz")` to change the method
+//! name to `xyz_foo` if the field is named `foo`. Note that an underscore is included by default,
+//! since Rust favors snake case here.
 //!
 //! ## Gotchas
 //!
@@ -188,6 +196,12 @@
 //!   names.
 //! - When defining a generic struct, you cannot use `VALUE` as a generic
 //!   parameter as this is what all setters are using.
+//!
+//! ## Debugging Info
+//!
+//! If you experience any problems during compilation, you can enable additional debug output
+//! by setting the environment variable `RUST_LOG=derive_builder=trace` before you call `cargo`
+//! or `rustc`. Example: `RUST_LOG=derive_builder=trace cargo test`.
 //!
 //! [builder pattern]: https://aturon.github.io/ownership/builders.html
 
@@ -203,13 +217,21 @@ extern crate env_logger;
 
 mod options;
 
+use std::borrow::Cow;
 use proc_macro::TokenStream;
-use options::{Options, SetterPattern};
+use quote::ToTokens;
+use options::{Options, OptionsBuilder, FieldMode, SetterPattern};
+use std::sync::atomic::{AtomicBool, Ordering, ATOMIC_BOOL_INIT};
+
+// beware: static muts are not threadsafe. :-)
+static mut LOGGER_INITIALIZED: AtomicBool = ATOMIC_BOOL_INIT; // false
 
 #[doc(hidden)]
-#[proc_macro_derive(Builder, attributes(setters, getters, setter, getter))]
+#[proc_macro_derive(Builder, attributes(setter))]
 pub fn derive(input: TokenStream) -> TokenStream {
-    env_logger::init().unwrap();
+    if unsafe { !LOGGER_INITIALIZED.compare_and_swap(false, true, Ordering::SeqCst) } {
+        env_logger::init().unwrap();
+    }
 
     let input = input.to_string();
 
@@ -249,13 +271,11 @@ fn filter_attr(attr: &&syn::Attribute) -> bool {
 
 fn builder_for_struct(ast: syn::MacroInput) -> quote::Tokens {
     debug!("Deriving Builder for '{}'.", ast.ident);
-    let opts = Options::from(ast.attrs);
+    let opts = Options::from(&ast.attrs);
     if !opts.setter_enabled() {
         trace!("Setters disabled for '{}'.", ast.ident);
         return quote!();
     }
-    debug!("Deriving Setters for '{}'.", ast.ident);
-    let setter_pattern = opts.setter_pattern();
 
     let fields = match ast.body {
         syn::Body::Struct(syn::VariantData::Struct(ref fields)) => fields,
@@ -265,46 +285,24 @@ fn builder_for_struct(ast: syn::MacroInput) -> quote::Tokens {
     let name = &ast.ident;
     let (impl_generics, ty_generics, where_clause) = ast.generics.split_for_impl();
     let funcs = fields.iter().map(|f| {
-        let f_name = &f.ident;
-        let ty = &f.ty;
 
-        trace!("Filtering field attributes");
-        let attrs = f.attrs.iter()
-            .filter(|a| {
-                let keep = filter_attr(a);
-                match keep {
-                    true => trace!("Keeping field attribute for setter {:?}", a),
-                    false => trace!("Ignoring field attribute {:?}", a)
-                }
-                keep
-            });
+        trace!("Parsing field {:?}.", f.ident.as_ref().map(|i| i.as_ref()).unwrap_or_default());
 
-        let vis = opts.setter_visibility();
-        debug!("Setter visibility = {:?}", vis);
+        let f_opts = OptionsBuilder::<FieldMode>::default()
+            .parse_attributes(&f.attrs)
+            .with_struct_options(&opts);
 
-        match *setter_pattern {
-            SetterPattern::Owned => quote!(
-                    #(#attrs)*
-                    #vis fn #f_name<VALUE: Into<#ty>>(self, value: VALUE) -> Self {
-                        let mut new = self;
-                        new.#f_name = value.into();
-                        new
-                }),
-            SetterPattern::Mutable => quote!(
-                    #(#attrs)*
-                    #vis fn #f_name<VALUE: Into<#ty>>(&mut self, value: VALUE) -> &mut Self {
-                        let mut new = self;
-                        new.#f_name = value.into();
-                        new
-                }),
-            SetterPattern::Immutable => quote!(
-                    #(#attrs)*
-                    #vis fn #f_name<VALUE: Into<#ty>>(&self, value: VALUE) -> Self {
-                        let mut new = self.clone();
-                        new.#f_name = value.into();
-                        new
-                }),
+        debug!("Setter prefix is {:?}", f_opts.setter_prefix());
+
+        let mut tokens = quote::Tokens::new();
+
+        if f_opts.setter_enabled() {
+            derive_setter(f, &f_opts).to_tokens(&mut tokens);
+        } else {
+            trace!("Skipping setter.");
         }
+
+        tokens
     });
 
     quote! {
@@ -313,4 +311,56 @@ fn builder_for_struct(ast: syn::MacroInput) -> quote::Tokens {
             #(#funcs)*
         }
     }
+}
+
+fn derive_setter(f: &syn::Field, opts: &Options) -> quote::Tokens {
+    trace!("Deriving setter.");
+    let ty = &f.ty;
+    let pattern = opts.setter_pattern();
+    let vis = opts.setter_visibility();
+    let fieldname = f.ident.as_ref().expect(&format!("Missing identifier for field {:?}.", f));
+    let funcname = if opts.setter_prefix().len() > 0 {
+        Cow::Owned(syn::Ident::new(format!("{}_{}", opts.setter_prefix(), fieldname)))
+    } else {
+        Cow::Borrowed(fieldname)
+    };
+
+    trace!("Filtering field attributes");
+    let attrs = f.attrs.iter()
+        .filter(|a| {
+            let keep = filter_attr(a);
+            match keep {
+                true => trace!("Keeping field attribute for setter {:?}", a),
+                false => trace!("Ignoring field attribute {:?}", a)
+            }
+            keep
+        });
+
+    let setter = match *pattern {
+        SetterPattern::Owned => quote!(
+                #(#attrs)*
+                #vis fn #funcname<VALUE: Into<#ty>>(self, value: VALUE) -> Self {
+                    let mut new = self;
+                    new.#fieldname = value.into();
+                    new
+            }),
+        SetterPattern::Mutable => quote!(
+                #(#attrs)*
+                #vis fn #funcname<VALUE: Into<#ty>>(&mut self, value: VALUE) -> &mut Self {
+                    let mut new = self;
+                    new.#fieldname = value.into();
+                    new
+            }),
+        SetterPattern::Immutable => quote!(
+                #(#attrs)*
+                #vis fn #funcname<VALUE: Into<#ty>>(&self, value: VALUE) -> Self {
+                    let mut new = self.clone();
+                    new.#fieldname = value.into();
+                    new
+            }),
+    };
+
+    debug!("Setter is {:?}", setter);
+
+    setter
 }